Add version info and cli version switch
Why you made the change:

I did this so that users would be able to identify the version of the command
line utility when reporting issues or just when they are curious. Also, added
the beginnings of a CHANGELOG.md file so that we can track changes and the
history of the project.