Add debugging, and strip filename for tests
When messing with a test file, and trying to find the non-test file we
need to strip out the _spec, or _test, etc... that way the actual file
can be found.

Normally this wouldn't be an issue, except that we didn't take into
account file names that were shorter than 5 characters or so. I've also
reduced the MIN_FILENAME_PERCENTAGE to offset this as well so it's not
as big of an issue - though that should have testing around it to make
sure it's still performant.