Modify suggested mapping to first write the file
Why you made the change:

I did this because in the scenario where you create a new file in vim you
technically have an unsaved buffer. Hence, if you switch away from that unsaved
buffer to say create its associated test file and then try and use AltCommand it
would say can't find alternate file. The reason being that it is looking at the
actual file system for the alternate file and it can't find it. Modifying the
mapping to first write the file resolves this issue. Therefore, I wanted to
include it as the default recommended mapping.