Remove the ruby scatter gather implementation
Why you made the change:

I did this because I created it just to benchmark it against the classic and the
results show that it isn't worth the overhead of creating the threads and doing
a scatter gather in Ruby. Also, the Rust version is significantly more
performant than the Ruby version so we will likely be moving away from the Ruby
version officially.