Update README.md to include About UpTech
Why you made the change:

I did this so people would be aware that UpTech Works, LLC was paying
for the development and maintenance of this project. Also, wanted to
include instructions on how to build from source for other platforms and
to include license information so people didn't have to go looking.