Fixed bug prevented Cucumber support from working
Why you made the change:

I did this so that Cucumber support would actually work properly.
Specifically, the issue was that because we were identifying anything
that matched ^features/ as a test file it was matching both
features/user_account.feature and
features/step_definitions/user_account_steps.rb as a test file, which in
turn was filtering the actual alternate out of the possible results.

So, I corrected the regex to be more explicit and only count paths
matching ^features/step_definitions/ as test files so that it wouldn't
count the features/user_accounts.feature as test file and would
appropriately find the alternate.