Extract is_test_file to module and add tests
Why you made the change:

I did this because I needed to add test coverage so that I could easily
and confidently make changes in the future. The extraction happened as a
side effect of the unit test coverage having to be so extensive as well
hints of future refactoring and restructuring coming.

I also added test coverage for the cleanse_path() function and the
get_filename_minus_extension() function so that I can refactor those
with confidence.