Extract strip_test_words into a module and add tests
Why you made the change:

I did this so that I would be able to refactor things with confidence as
I have plans to refactor and add some more features. Anyways, through
just adding the test coverage I already identified a scenario that was
missed, "Ruby Rake Minitest" so I added the test for that and added
support for that in this commit as well.