Generalize extension handling & add groovy support
Why you made the change:

I generalized extension handling into matching a . character following
by one ore more word characters at the end of the string. This should
make alt work out of the box with more languages and tools.

I also added the Spec option at the end of a file stem to be stripped
before comparing for similarity scoring. I did this because a number of
testing tools use a naming scheme where the file stem ends in Spec. One
is specifically Groovy. So, this adds support for Groovy tests as well
as handles other unnamed tools and test frameworks.