Changed args to proper args & reference installed
I did this so that it will behave correctly in terms since it is now
getting the correct command line arguments in the benchmarks.

I also set it up so that one of the sets references the "Rust Classic"
or rather the version currently installed at /usr/local/bin/alt and the
"Latest Build" version as the one that is relatively location at
./target/release/alt. This makes it easier to benchmark the changes you
just made against the latest stable release in most cases.