Add full blown scenarios to benchmark.rb script
I did this so that we would not only get benchmarks on the scenario
where alt is handed the possibles but also the full blown scenarios
where someone uses find combined with alt and the scenario where someone
uses the Glob/WalkDir path with alt. That way we could see the
performance between those two use cases as they are the primary use
cases.