Switch from Walkdir to ignore::WalkBuilder
I did this because the ignore crate is the crate that ripgrep uses and
is built on top of Walkdir. It supports filtering and ignoring walking
directories based on globs defined in .ignore, .gitignore (in project
and global), hidden files, etc. This gives us the performance gains of
Walkdir while still being able to filter out things you don't want to
include as possible alternates. All without having to use find or some
other command and piping the output into alt. Which in the end also
saves us the cost of starting up an extra program. This should be a
robust enough solution that we can even consider getting rid of the the
-f option, maybe.

I also upgraded the dependencies we use. The only change I had to make
in terms of the dependency upgrade was the change to
alt/path/alteration/mod.rs because the API for the regex crate changed
slightly.