Add multiple alterantes & rework scoring
Add multiple alterantes so that users don't have to be limited to just
one alternate. Instead they can be presented with a number of alternates
and then select which one they want if any.

As part of this ran into a couple bug in how we were scoring and how we
were previously sorting the scored paths. To address these issues I
revisited the original concept of just using scores and got rid of the
pre-filter stuff and test file detection, etc. All of it was brittle and
made maintenance of this tool a pain in the but, not to mention less of
a valuable general tool.

To make it more valuable I changed the path scoring to distinguish
between the filename portion of the path and the rest of the path. They
are both scored independently and then the filename portion is weighted
higher. I did this because I believe it is true that generally if a
filename has similarity it is more significant than if the directories
have similarities but the filenames don't.

I also cleaned up the entire code base in terms of organizing around
concerns, added a bunch of test coverage, etc. In general just tried to
make it more maintainable and better overall in the long term.

[changelog]
added: multi-line/multi-alternate output
changed: default behavior to output all possible alternates sorted by score
added: CLI option to truncate possible alterates down to a specific number
changed: scoring algorithm to be able to weight filenames different than paths
added: CLI option to override the scoring algorithms default filename weight
added: CLI option to override the scoring algorithms default path weight
added: support for optionally using threads to score in parallel
added: CLI option to enable parallel scoring
fixed: bug where scored paths weren't being sorted properly
fixed: bug where some paths would score as NaN and make sorting off
changed: Vim example to a NeoVim Lua example as that is what I use