Add leading_edge_filename scoring to path scoring
Add leading_edge_filename scoring to path scoring so that when paths are
scored they are scored. I did this because the current scoring algorithm
would score paths that don't match the leading edge filename over ones
that did match the leading edge filename. This change makes it so we can
score paths that match the leading edge filename over paths that don't
if we want.

A user can have the same behavior as before by passing a weight of 0 for
the leading edge filename weight.

skip_ticket_id_check

[changelog]
changed: path scoring to add scoring of leading_edge_filename

<!-- ps-id: 24a7ab4c-3785-43b8-ac80-284b23f9288d -->